Terminado Aula17 e Adicionado Aula18

+++ eclipse-workspace/calculadora_api/src/main/java/br/com/db1/calculadoraapi/calculadora_api/model/OperacaoMatematica.java
@@ -3,29 +3,29 @@ package br.com.db1.calculadoraapi.calculadora_api.model;
 public enum OperacaoMatematica {
 	SOMA {
 		@Override
-		public Integer executar(Integer num1, Integer num2) {
+		public Double executar(Double num1, Double num2) {
 			return num1 + num2;
 		}
 	},
 	SUBTRACAO {
 		@Override
-		public Integer executar(Integer num1, Integer num2) {
+		public Double executar(Double num1, Double num2) {
 			return num1 - num2;
 		}
 	},
 	DIVISAO {
 		@Override
-		public Integer executar(Integer num1, Integer num2) {
+		public Double executar(Double num1, Double num2) {
 			return num1 / num2;
 		}
 	},
 	MULTIPLICACAO {
 		@Override
-		public Integer executar(Integer num1, Integer num2) {
+		public Double executar(Double num1, Double num2) {
 			return num1 * num2;
 		}
 	};
 
-	public abstract Integer executar(Integer num1, Integer num2);
+	public abstract Double executar(Double num1, Double num2);
 
 }
